# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API do Desafio Final de backend do PD Itabira (Protegida)",
    "description": "Gestão de Alunos, Cursos e Matrículas usando Identificadores Amigáveis. UUIDs ocultos por segurança.",
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor Local"
    }
  ],
  "components": {
    "schemas": {
      "ErroPadrao": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "statusCode": { "type": "integer" }
        }
      }
    }
  },
  "paths": {
    "/alunos": {
      "get": {
        "tags": ["Alunos"],
        "summary": "Lista alunos (filtra por nome ou código amigável)",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "description": "Parte do nome",
            "schema": { "type": "string" }
          },
          {
            "name": "codigo_aluno",
            "in": "query",
            "description": "Número da matrícula do aluno",
            "schema": { "type": "integer" }
          }
        ],
        "responses": { "200": { "description": "Sucesso" } }
      },
      "post": {
        "tags": ["Alunos"],
        "summary": "Cadastra um novo aluno",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nome", "email"],
                "properties": {
                  "nome": { "type": "string" },
                  "email": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Aluno criado" } }
      }
    },
    "/alunos/{codigo}": {
      "get": {
        "tags": ["Alunos"],
        "summary": "Busca aluno pelo código amigável",
        "parameters": [{ "name": "codigo", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Sucesso" } }
      },
      "put": {
        "tags": ["Alunos"],
        "summary": "Atualiza aluno pelo código",
        "parameters": [{ "name": "codigo", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "nome": { "type": "string" }, "email": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Atualizado" } }
      },
      "delete": {
        "tags": ["Alunos"],
        "summary": "Remove aluno pelo código",
        "parameters": [{ "name": "codigo", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Removido" } }
      }
    },
    "/cursos": {
      "get": {
        "tags": ["Cursos"],
        "summary": "Lista cursos (filtra por nome ou código)",
        "parameters": [
          { "name": "nome", "in": "query", "schema": { "type": "string" } },
          { "name": "codigo_curso", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Sucesso" } }
      },
      "post": {
        "tags": ["Cursos"],
        "summary": "Cria um curso",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "required": ["nome_curso"], "properties": { "nome_curso": { "type": "string" }, "carga_horaria": { "type": "integer" } } } } }
        },
        "responses": { "201": { "description": "Criado" } }
      }
    },
    "/cursos/{codigo}": {
      "put": {
        "tags": ["Cursos"],
        "summary": "Atualiza curso pelo código",
        "parameters": [{ "name": "codigo", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "nome_curso": { "type": "string" }, "carga_horaria": { "type": "integer" } } } } }
        },
        "responses": { "200": { "description": "Atualizado" } }
      },
      "delete": {
        "tags": ["Cursos"],
        "summary": "Deleta curso pelo código",
        "parameters": [{ "name": "codigo", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Removido" } }
      }
    },
    "/matriculas": {
      "get": {
        "tags": ["Matrículas"],
        "summary": "Lista todas as matrículas (Usa apenas códigos amigáveis)",
        "parameters": [{ "name": "codigo_matricula", "in": "query", "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Sucesso" } }
      },
      "post": {
        "tags": ["Matrículas"],
        "summary": "Matricular aluno usando códigos curtos",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["codigo_aluno", "codigo_curso"],
                "properties": {
                  "codigo_aluno": { "type": "integer" },
                  "codigo_curso": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Matriculado" } }
      }
    },
    "/matriculas/{codigo}/status": {
      "patch": {
        "tags": ["Matrículas"],
        "summary": "Alterar status pelo código da matrícula",
        "parameters": [{ "name": "codigo", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string", "enum": ["ativa", "cancelada", "concluida"] } } } } }
        },
        "responses": { "200": { "description": "Alterado" } }
      }
    }
  }
}